Introduce SchemaValidator, validation of table schemas in plugin service
