+++ test/plugin/cassandraSchema/SchemaValidator.js
@@ -1,10 +1,10 @@
 const assert = require('assert');
 
-const SchemaCreator = require('../../../plugin/cassandraSchema/SchemaCreator');
+const SchemaValidator = require('../../../plugin/cassandraSchema/SchemaValidator');
 
 describe('Schema Creator', () => {
     it('Should return error if parameters field is frozen list type', () => {
-        const err = SchemaCreator.getSchemaError({
+        const err = SchemaValidator.getSchemaError({
             parameters: 'frozen<list<int>>'
         });
 
@@ -12,7 +12,7 @@ describe('Schema Creator', () => {
     });
 
     it('Should return error if parameters field is tuple type', () => {
-        const err = SchemaCreator.getSchemaError({
+        const err = SchemaValidator.getSchemaError({
             parameters: 'tuple<double>'
         });
 
@@ -20,7 +20,7 @@ describe('Schema Creator', () => {
     });
 
     it('Should return error if parameters field is frozen set type', () => {
-        const err = SchemaCreator.getSchemaError({
+        const err = SchemaValidator.getSchemaError({
             parameters: 'frozen<set<int>>'
         });
 
@@ -28,7 +28,7 @@ describe('Schema Creator', () => {
     });
 
     it('Should return error if parameters field is timestamp', () => {
-        const err = SchemaCreator.getSchemaError({
+        const err = SchemaValidator.getSchemaError({
             parameters: 'timestamp'
         });
 
@@ -36,7 +36,7 @@ describe('Schema Creator', () => {
     });
 
     it('Should return error if parameters field is map with at least one non string component (key or value)', () => {
-        const err = SchemaCreator.getSchemaError({
+        const err = SchemaValidator.getSchemaError({
             parameters: 'map<int,text>'
         });
 
@@ -44,7 +44,7 @@ describe('Schema Creator', () => {
     });
 
     it('Should return error if parameters field is built in basic type and not a string', () => {
-        const err = SchemaCreator.getSchemaError({
+        const err = SchemaValidator.getSchemaError({
             parameters: 'int'
         });
 
@@ -52,7 +52,7 @@ describe('Schema Creator', () => {
     });
 
     it('Should return null if parameters field is text type', () => {
-        const err = SchemaCreator.getSchemaError({
+        const err = SchemaValidator.getSchemaError({
             parameters: 'text'
         });
 
@@ -60,7 +60,7 @@ describe('Schema Creator', () => {
     });
 
     it('Should return null if parameters field is ascii type', () => {
-        const err = SchemaCreator.getSchemaError({
+        const err = SchemaValidator.getSchemaError({
             parameters: 'ascii'
         });
 
@@ -68,7 +68,7 @@ describe('Schema Creator', () => {
     });
 
     it('Should return null if parameters field is varchar type', () => {
-        const err = SchemaCreator.getSchemaError({
+        const err = SchemaValidator.getSchemaError({
             parameters: 'varchar'
         });
 
@@ -76,7 +76,7 @@ describe('Schema Creator', () => {
     });
 
     it('Should return null if parameters field is frozen<map<text,text>> type', () => {
-        const err = SchemaCreator.getSchemaError({
+        const err = SchemaValidator.getSchemaError({
             parameters: 'frozen<map<text,text>>'
         });
 
@@ -84,7 +84,7 @@ describe('Schema Creator', () => {
     });
 
     it('Should return null if parameters field is frozen<map<text,text>> type', () => {
-        const err = SchemaCreator.getSchemaError({
+        const err = SchemaValidator.getSchemaError({
             parameters: 'frozen<map<text,text>>'
         });
 
@@ -92,7 +92,7 @@ describe('Schema Creator', () => {
     });
 
     it('Should return null if parameters field is UDT with frozen<>', () => {
-        const err = SchemaCreator.getSchemaError({
+        const err = SchemaValidator.getSchemaError({
             parameters: 'frozen<some_udt_with_int_>'
         });
 
@@ -100,7 +100,7 @@ describe('Schema Creator', () => {
     });
 
     it('Should return null if parameters field is UDT without frozen<>', () => {
-        const err = SchemaCreator.getSchemaError({
+        const err = SchemaValidator.getSchemaError({
             parameters: 'int_double_udt_detection_test'
         });
 
@@ -117,7 +117,7 @@ describe('Schema Creator', () => {
             }
         };
 
-        const errors = SchemaCreator.getSchemasErrors(tables);
+        const errors = SchemaValidator.getSchemasErrors(tables);
 
         assert.equal(errors.length, 0);
     });
